tests/wpt/tests/long-animation-frame/tentative/loaf-trusted-types.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Long Animation Frame Timing: basic</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<h1>Long Animation Frame: eval</h1>
<div id="log"></div>
<script id="element"></script>
<script>
setup(() =>
  assert_implements("trustedTypes" in window,
    'trustedTypes are not supported.'));

/* This invokes the createHTML trusted type */
trustedTypes.defaultPolicy = trustedTypes.createPolicy("default", {
    createHTML: (input) => input
});

document.getElementById("element").innerHTML = `
  busy_wait(very_long_frame_duration / 2);
  Promise.resolve().then(() => {
    busy_wait(very_long_frame_duration / 2);
  })
`;

promise_test(async t => {
  await new Promise(resolve => new PerformanceObserver(entries => {
    const entry = entries.getEntries()[0];
    if (entry.duration >= (very_long_frame_duration - 5) &&
      entry.scripts.some(script => script.duration >= very_long_frame_duration)) {
        resolve(entry);
      }
  }).observe({type: "long-animation-frame", buffered: true}));
});

</script>
</body>
